+++ traffic_ops/traffic_ops_golang/api/generic_crud.go
@@ -23,11 +23,12 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"net/http"
+	"time"
+
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-rfc"
 	ims "github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/util/ims"
-	"net/http"
-	"time"
 
 	"github.com/apache/trafficcontrol/lib/go-tc"
 	"github.com/apache/trafficcontrol/lib/go-util"
@@ -56,6 +57,7 @@ type GenericUpdater interface {
 	APIInfo() *APIInfo
 	SetLastUpdated(tc.TimeNoMod)
 	UpdateQuery() string
+	GetLastUpdated() (*time.Time, bool, error)
 }
 
 type GenericDeleter interface {
@@ -223,7 +225,18 @@ func GenericRead(h http.Header, val GenericReader, useIMS bool) ([]interface{}, 
 }
 
 // GenericUpdate handles the common update case, where the update returns the new last_modified time.
-func GenericUpdate(val GenericUpdater) (error, error, int) {
+func GenericUpdate(h http.Header, val GenericUpdater) (error, error, int) {
+	existingLastUpdated, found, err := val.GetLastUpdated()
+	if err == nil && found == false {
+		return errors.New("no " + val.GetType() + " found with this id"), nil, http.StatusNotFound
+	}
+	if err != nil {
+		return nil, err, http.StatusInternalServerError
+	}
+	if !IsUnmodified(h, *existingLastUpdated) {
+		return errors.New("resource was modified"), nil, http.StatusPreconditionFailed
+	}
+
 	rows, err := val.APIInfo().Tx.NamedQuery(val.UpdateQuery(), val)
 	if err != nil {
 		return ParseDBError(err)