+++ traffic_ops/traffic_ops_golang/federations/federations.go
@@ -115,9 +115,7 @@ func Get(w http.ResponseWriter, r *http.Request) {
 	}
 	allFederations := addResolvers([]tc.IAllFederation{}, feds, fedsResolvers)
 	if maxTime != nil && api.SetLastModifiedHeader(r, useIMS) {
-		// RFC1123
-		date := maxTime.Format("Mon, 02 Jan 2006 15:04:05 MST")
-		w.Header().Add(rfc.LastModified, date)
+		api.AddLastModifiedHdr(w, *maxTime)
 	}
 	api.WriteResp(w, r, allFederations)
 }