+++ traffic_ops/v3-client/tenant.go
@@ -119,20 +119,25 @@ func (to *Session) CreateTenant(t *tc.Tenant) (*tc.TenantResponse, error) {
 	return &data, nil
 }
 
-// UpdateTenant updates the Tenant matching the ID it's passed with
-// the Tenant it is passed.
-func (to *Session) UpdateTenant(id string, t *tc.Tenant) (*tc.TenantResponse, error) {
+func (to *Session) UpdateTenantWithHdr(id string, t *tc.Tenant, header http.Header) (*tc.TenantResponse, error) {
 	var data tc.TenantResponse
 	jsonReq, err := json.Marshal(t)
 	if err != nil {
 		return nil, err
 	}
-	_, err = put(to, fmt.Sprintf(API_TENANT_ID, id), jsonReq, &data)
+	_, err = put(to, fmt.Sprintf(API_TENANT_ID, id), jsonReq, &data, header)
 	if err != nil {
 		return nil, err
 	}
 
 	return &data, nil
+}
+
+// UpdateTenant updates the Tenant matching the ID it's passed with
+// the Tenant it is passed.
+// Deprecated: UpdateTenant will be removed in 6.0. Use UpdateTenantWithHdr.
+func (to *Session) UpdateTenant(id string, t *tc.Tenant) (*tc.TenantResponse, error) {
+	return to.UpdateTenantWithHdr(id, t, nil)
 }
 
 // DeleteTenant deletes the Tenant matching the ID it's passed.