+++ traffic_portal/app/src/common/modules/form/deliveryService/edit/FormEditDeliveryServiceController.js
@@ -149,8 +149,8 @@ var FormEditDeliveryServiceController = function(deliveryService, origin, topolo
 					if (autoFulfilled) {
 						// assign the ds request
 						promises.push(deliveryServiceRequestService.assignDeliveryServiceRequest(response.id, userModel.user.id));
-						// set the status to 'complete'
-						promises.push(deliveryServiceRequestService.updateDeliveryServiceRequestStatus(response.id, 'complete'));
+						// set the status to 'submitted'
+						promises.push(deliveryServiceRequestService.updateDeliveryServiceRequestStatus(response.id, 'submitted'));
 					}
 				}
 			);
@@ -234,10 +234,20 @@ var FormEditDeliveryServiceController = function(deliveryService, origin, topolo
 							deliveryServiceService.updateDeliveryService(deliveryService).
 								then(
 									function() {
+										// upon successful update of ds, set the dsr to 'complete'
+										deliveryServiceRequestService.getDeliveryServiceRequests({ xmlId: deliveryService.xmlId, status: 'submitted' }).then(
+											function(response) {
+												deliveryServiceRequestService.updateDeliveryServiceRequestStatus(response[0].id, 'complete');
+											}
+										);
 										$state.reload(); // reloads all the resolves for the view
 										messageModel.setMessages([ { level: 'success', text: 'Delivery Service [ ' + deliveryService.xmlId + ' ] updated' } ], false);
 									}
-								);
+								).catch(function(fault) {
+									// if the ds update fails, send to dsr view w/ error message
+									locationUtils.navigateToPath('/delivery-service-requests');
+									messageModel.setMessages(fault.data.alerts, true);
+							});
 						}).catch(function(fault) {
 							$anchorScroll(); // scrolls window to top
 							messageModel.setMessages(fault.data.alerts, false);