+++ traffic_portal/app/src/common/modules/table/servers/TableServersController.js
@@ -305,11 +305,6 @@ var TableServersController = function(tableName, servers, filter, $scope, $state
 			sshCellRenderer: SSHCellRenderer,
 			updateCellRenderer: UpdateCellRenderer
 		},
-		onRowDoubleClicked: function(params) {
-			locationUtils.navigateToPath('/servers/' + params.data.id);
-			// Event is outside the digest cycle, so we need to trigger one.
-			$scope.$apply();
-		},
 		columnDefs: columns,
 		enableCellTextSelection:true,
 		suppressMenuHide: true,
@@ -325,12 +320,24 @@ var TableServersController = function(tableName, servers, filter, $scope, $state
 		pagination: true,
 		paginationPageSize: $scope.pageSize,
 		rowBuffer: 0,
-		onColumnResized: function(params) {
-			localStorage.setItem(tableName + "_table_columns", JSON.stringify($scope.gridOptions.columnApi.getColumnState()));
-		},
 		tooltipShowDelay: 500,
 		allowContextMenuWithControlKey: true,
 		preventDefaultOnContextMenu: true,
+		colResizeDefault: "shift",
+		onColumnVisible: function(params) {
+			if (params.visible){
+				return;
+			}
+		    for (let column of params.columns) {
+				if (column.filterActive) {
+					const filterModel = $scope.gridOptions.api.getFilterModel();
+					if (column.colId in filterModel) {
+						delete filterModel[column.colId];
+						$scope.gridOptions.api.setFilterModel(filterModel);
+					}
+				}
+			}
+		},
 		onCellContextMenu: function(params) {
 			$scope.showMenu = true;
 			$scope.menuStyle.left = String(params.event.clientX) + "px";
@@ -351,7 +358,15 @@ var TableServersController = function(tableName, servers, filter, $scope, $state
 			$scope.server = params.data;
 			$scope.$apply();
 		},
-		colResizeDefault: "shift"
+		onRowDoubleClicked: function(params) {
+			locationUtils.navigateToPath('/servers/' + params.data.id);
+			// Event is outside the digest cycle, so we need to trigger one.
+			$scope.$apply();
+		},
+		onColumnResized: function(params) {
+			localStorage.setItem(tableName + "_table_columns", JSON.stringify($scope.gridOptions.columnApi.getColumnState()));
+		},
+
 	};
 
 	/** These three functions are used by the context menu to determine what functionality to provide for a server. */
@@ -394,7 +409,7 @@ var TableServersController = function(tableName, servers, filter, $scope, $state
 			fileName: "servers.csv",
 		};
 		$scope.gridOptions.api.exportDataAsCsv(params);
-	}
+	};
 
 	/**** Context menu functions ****/
 
@@ -554,7 +569,11 @@ var TableServersController = function(tableName, servers, filter, $scope, $state
 		localStorage.setItem(tableName + "_page_size", value);
 	};
 
-	$scope.clearColFilters = function() {
+	$scope.clearTableFilters = function() {
+		// clear the quick search
+		$scope.quickSearch = '';
+		$scope.onQuickSearchChanged();
+		// clear any column filters
 		$scope.gridOptions.api.setFilterModel(null);
 	};